fix google cast not working
